Move logging function & call to rbac/middleware to avoid logging request twice
